requirements.txt
apache-airflow>=2.7.0
python-dotenv
requests
soda-core-postgres==3.3.14
pytest==8.3.2
